Refactor - reuse UploadButton, move icons
