+++ src/components/ReactFileUtilities/icons/AttachmentIcon.tsx
