+++ src/components/ReactFileUtilities/icons/CloseIcon.tsx
