+++ src/components/ReactFileUtilities/icons/PictureIcon.tsx
