+++ src/components/ReactFileUtilities/icons/index.ts
@@ -1,0 +1,5 @@
+export * from './AttachmentIcon';
+export * from './CloseIcon';
+export * from './FilePlaceholderIcon';
+export * from './PictureIcon';
+export * from './RetryIcon';
